Atlas: Ported terrain texture selection panel to JS. Removed textual labels on each texture. wxJS: Added incomplete wxNotebook. Added wxWindow.toolTip.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5174 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/GameInterface/Brushes.cpp
@@ -66,7 +66,7 @@ void Brush::SetData(int w, int h, const std::vector<float>& data)
 
 	m_Data = data;
 	
-	debug_assert(data.size() == w*h);
+	debug_assert(data.size() == (size_t)(w*h));
 }
 
 void Brush::GetCentre(int& x, int& y) const
